projects/py-app1/pyproject.toml
[project]
name = "py-app1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastapi[standard]>=0.135.1",
  "py-lib1",
  "python-dotenv>=1.2.2",
]

[tool.uv.sources]
py-lib1 = { workspace = true }
